QOL improvements from my independently developed home version
During the NEA, I had a codebase similar to the one I was developing during the 20 hours. I was able to experiment with cleaner practises, without having to spend most of my time writing up tests.

One of these cleaner practises was to introduce a set of interfaces and associated convenience functionality - another was a different set of packages names.